refactor: extract shared JSON-RPC endpoint core into rpc.ts
client.ts and broker-client.ts each hand-maintained a copy of the
wire-level plumbing (pending-request tracking, response/error/server-
request/notification dispatch, newline-buffered parsing, handler
registries); several #10 review findings were drift between the copies.

createRpcEndpoint() now owns all of it, parameterized only by what
genuinely differs per transport: the raw send, a liveness probe
(downReason), overflow teardown, and the reason strings tests pin down.
client.ts keeps spawn/stderr/exit/platform-shutdown/handshake;
broker-client.ts keeps socket connect/close/handshake. Protocol helpers
(parseMessage, formatters, type guards) moved to rpc.ts with re-exports
from client.ts so broker-server.ts and tests are untouched.

Behavioral deltas (all benign): a buffer overflow now fires onClose
handlers immediately instead of waiting for process exit; feed() goes
inert after overflow instead of re-triggering per chunk; post-failure
requests reject with the original failure reason rather than a generic
'Broker connection closed'.
